+++ lib/ui/src/modules/ui/components/left_panel/stories/index.test.js
@@ -1,7 +1,7 @@
 import { shallow, mount } from 'enzyme';
 import React from 'react';
-import Stories from './stories';
-import { createHierarchy } from '../../libs/hierarchy';
+import Stories from './index';
+import { createHierarchy } from '../../../libs/hierarchy';
 
 describe('manager.ui.components.left_panel.stories', () => {
   describe('render', () => {
