+++ lib/ui/src/modules/ui/components/left_panel/stories/tree_node_type.js
@@ -1,0 +1,5 @@
+export default {
+  NAMESPACE: 'namespace',
+  COMPONENT: 'component',
+  STORY: 'story',
+};
